Moved Book object to separate file, functionized parsing.
