+++ main.py
@@ -1,22 +1,12 @@
 from bs4 import BeautifulSoup
-import os, time, glob
+from models import Book
+import os
+import glob
+import time
+
 
 start = time.time()
-results = []
 os.chdir("data")
-
-
-class Book:
-    def __init__(self, title, author, price, ISBN, weight):
-        self.title = title
-        self.author = author
-        self.price = price
-        self.ISBN = ISBN
-        self.weight = weight
-
-
-    def __repr__(self):
-        return "{1} by {2}".format(self.title, self.author)
 
 
 def extract_content(tag):
@@ -24,33 +14,41 @@ def extract_content(tag):
     end = tag.find("<", start)
     return tag[start:end].strip()
 
-for file in glob.glob("*.html"):
-    working_file = open(file, "r", encoding="ISO-8859-1")
-    soup = BeautifulSoup(working_file.read())
 
-    title = extract_content(str(soup.find(id="btAsinTitle")))
+def parse_files(directory):
+    results = []
+    for file in directory:
+        working_file = open(file, "r", encoding="ISO-8859-1")
+        soup = BeautifulSoup(working_file.read())
 
-    author_data = str(soup.findAll("meta", {"name": "keywords"})[0])
-    author = author_data[author_data.find('"') + 1:author_data.find(",")]
+        title = extract_content(str(soup.find(id="btAsinTitle")))
 
-    price_data = soup.findAll("span", {"class": "bb_price"})[0]
-    price = extract_content(str(price_data))
+        author_data = str(soup.findAll("meta", {"name": "keywords"})[0])
+        author = author_data[author_data.find('"') + 1:author_data.find(",")]
 
-    remaining_data = str(soup.findAll("td", {"class": "bucket"})[0])
-    soup = BeautifulSoup(remaining_data)
-    remaining_data = soup.findAll("li")
+        price_data = soup.findAll("span", {"class": "bb_price"})[0]
+        price = extract_content(str(price_data))
 
-    for data in remaining_data:
-        if "ISBN-10" in str(data):
-            ISBN = str(data)
-        elif "Weight" in str(data):
-            weight = str(data)
+        remaining_data = str(soup.findAll("td", {"class": "bucket"})[0])
+        soup = BeautifulSoup(remaining_data)
+        remaining_data = soup.findAll("li")
 
-    ISBN = ISBN[ISBN.find("</b> ") + 5:ISBN.find("</li>")]
-    weight = weight[weight.find("</b> ") + 5:weight.find(" (")]
+        for data in remaining_data:
+            if "ISBN-10" in str(data):
+                ISBN = str(data)
+            elif "Weight" in str(data):
+                weight = str(data)
 
-    new_book = Book(title, author, price, ISBN, weight)
-    results.append(new_book)
+        ISBN = ISBN[ISBN.find("</b> ") + 5:ISBN.find("</li>")]
+        weight = weight[weight.find("</b> ") + 5:weight.find(" (")]
 
+        new_book = Book(title, author, price, ISBN, weight)
+        results.append(new_book)
+
+    return results
+
+
+directory = glob.glob("*.html")
+results = parse_files(directory)
 print(results)
 print(time.time() - start)
